Remove vSphere instance management
Previous to this commit, if Rgbank::Infrastructure::Web components were
defined in the environment catalog, vsphere VMs would be ensured to
exist for each Rgbank::Infrstructure::Web component.  However, this
stragey of provisioning VMs as part of an application deployment only
worked with VMs where the client certname could be defined upfront. This
will not be the way we manage virtual infrastructure in the future, so
it doesn't make sense to have this in the module right now.

This commit removes the the ability to manage virtual infrastructure as
part of an app deployment, which has the added benefit of significantly
cleaning up the code.