Assume only one DB and Load component resource
Previous to this commit, an array of load balancer and DB components
were generated and iterated over, with the obscure assumption there
would only be one iteration.  This was horribly confusing to read.

This commit moves the assumption to collection of the components, only
returning the first element of the array. This means only one component
resource has to be declared.  Much cleaner