Updating web::base for github source
Prior to this commit the vcsrepo would checkout the rgbank git repo
into the theme directory.  The problem was that there are
subdirectories that are not present in the Jenkins built artifacts.
This commit creates a staging area to checkout git to and then sets
a symlink in the themes directory that points to the git repo src
folder